% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Download wildfinder dataset}
\usage{
dl_wildfinder_data(i, overwrite = FALSE)
}
\arguments{
\item{i}{select the csv file to download}

\item{overwrite}{a \code{logical}. If \code{TRUE}, the file will be downloaded again
and the previous version will be replaced.}
}
\value{
No return value.
}
\description{
This function downloads the wildfinder datasets (csv file) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}. The file
won't be downloaded if already exists locally (except if \code{overwrite = TRUE}).
The files \code{wildfinder-ecoregions_list.csv}, 'wildfinder-ecoregions_species.csv'
and 'wildfinder-mammals_list.csv' will be stored in
\verb{data/wildfinder/}. This folder will be created if required.
}
